# Row can update separately from rest of SortableTable